pyproject.toml
[project]
name = "lottery-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Grigori Kartashyan",email = "grigori.kartashyan@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
